Users can only edit and delete own posts

+++ app/controllers/wikis_controller.rb
@@ -1,63 +1,72 @@
 class WikisController < ApplicationController
+
+  before_filter :authenticate_user!, except: [ :index, :show ]
+
+  after_action :verify_authorized, :except => :index
+
   def index
     @wikis = Wiki.all
   end
 
   def show
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def new
     @wiki = Wiki.new
+    authorize @wiki
+  end
+
+  def create
+    @wiki = Wiki.create(wiki_params)
+    @wiki.user = current_user
+    authorize @wiki
+
+    if @wiki.save
+      redirect_to @wiki, notice: "Wiki was saved successfully."
+    else
+      flash[:error] = "Error creating wiki. Please try again."
+    end
   end
 
   def edit
-     @wiki = Wiki.find(params[:id])
+    @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
-  
-  def update
-     @wiki = Wiki.find(params[:id])
-     @wiki.title = params[:wiki][:title]
-     @wiki.body = params[:wiki][:body]
- 
-     if @wiki.save
-       flash[:notice] = "Wiki was updated."
-       redirect_to @wiki
-     else
-       flash.now[:alert] = "There was an error saving the wiki. Please try again."
-       render :edit
-     end
-  end
-  
-  def create
-     @wiki = Wiki.new
-     @wiki.title = params[:wiki][:title]
-     @wiki.body = params[:wiki][:body]
 
-     if @wiki.save
-       flash[:notice] = "Wiki was saved."
-       redirect_to @wiki
-     else
-       flash.now[:alert] = "There was an error saving the wiki. Please try again."
-       render :new
-     end
+  def update
+    @wiki = Wiki.find(params[:id])
+    authorize @wiki
+
+    @wiki.assign_attributes(wiki_params)
+
+    if @wiki.save
+      flash[:notice] = "Wiki was updated."
+      redirect_to @wiki
+    else
+      flash[:error] = "Error saving wiki. Please try again."
+      render :edit
+    end
   end
-  
+
   def destroy
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
 
     if @wiki.destroy
       flash[:notice] = "\"#{@wiki.title}\" was deleted successfully."
-      redirect_to @wiki
+      redirect_to action: :index
     else
       flash[:error] = "There was an error deleting the wiki."
       render :show
     end
   end
-  
-  private
+
+ private
 
   def wiki_params
     params.require(:wiki).permit(:title, :body, :public)
   end
+
 end
